Restructured directories and added resources
