--- CardPage.xaml
@@ -1,19 +1,0 @@
-<ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="Combat_Critters_2._0.CardPage"
-             Title="Cards">
-
-
-    <ScrollView>
-        <FlexLayout
-            Direction="Row"
-            Wrap="Wrap"
-            JustifyContent="Center"
-            AlignItems="Center"
-            Padding="10">
-
-            <!-- Card Frame -->
-
-        </FlexLayout>
-    </ScrollView>
-</ContentPage>
