--- CardPage.xaml.cs
@@ -1,13 +1,0 @@
-
-using Microsoft.Maui.Controls;
-
-namespace Combat_Critters_2._0;
-
-public partial class CardPage : ContentPage
-{
-    public CardPage()
-    {
-        InitializeComponent();
-    }
-
-}
