+++ Components/Card.xaml.cs
@@ -1,0 +1,11 @@
+namespace Combat_Critters_2._0.Components
+{
+    public partial class Card : ContentView
+    {
+        public Card()
+        {
+            InitializeComponent();
+        }
+    }
+
+}
