+++ Pages/CardPage.xaml.cs
@@ -1,0 +1,14 @@
+
+
+namespace Combat_Critters_2._0.Pages
+{
+    public partial class CardPage : ContentPage
+    {
+        public CardPage()
+        {
+            InitializeComponent();
+        }
+
+    }
+}
+
