+++ Pages/CreateAccountPage.xaml
@@ -1,7 +1,7 @@
 <ContentPage
     xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
     xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-    x:Class="Combat_Critters_2._0.CreateAccountPage"
+    x:Class="Combat_Critters_2._0.Pages.CreateAccountPage"
     Title="Create Account">
 
     <!-- Grid Layout -->
