+++ Pages/DeckPage.xaml
@@ -1,6 +1,6 @@
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="Combat_Critters_2._0.DeckPage"
+             x:Class="Combat_Critters_2._0.Pages.DeckPage"
              Title="Decks">
