+++ Pages/DeckPage.xaml.cs
@@ -1,4 +1,4 @@
-namespace Combat_Critters_2._0;
+namespace Combat_Critters_2._0.Pages;
 
 public partial class DeckPage : ContentPage
 {
